Added an unconventional read/write function.
The new function exported to the lua library allows one to read and write a single byte value in a signed and floating point scaled way. (might not be as useful as I hope, maybe doing the same with a 4 byte int would be a better idea.) This allows a way to link game variables together with a smooth transistion between them.