Added prerequisite for a transport triggered architecture.
Added a page type (CallbackPage) that just returns zeroes when read and calls a function on write.

Also added a protected member type (VMS_POINTER) to be used in the future for storing addresses into the virtual memory spaces.

I am currently pondering if I should make the type a struct that stores the memory space it is created from -
 so it can prevent trying to use an address from one memory space as a pointer into a different memory space.