(untested) Changed page visualization to be more portable.
The lua memory mapper should now display a window every time a page is linked to a memory space.
Changed the page window manager class to manage page visuialization classes. call "addVisualizer" with a pointer to the address space to add it to the list of windows. they are automatically removed from that list when the window is closed.