Made LuaMap and LuaMap32 report build info.
Windows builds of LuaMap and LuaMap32 now report their github revision and build time on every build.
This is done by using codeblocks pre-build step to generate a "revinfo.h" file that is included into the "LuaMemoryMapper.cpp" file if debug messages are switched on in the "config.h" file.