repairing and updating lua builds
I was using the version of lua from the game engine "LOVE" so I could test the library in-situ.
Now I have updated to lua5.4.4.
I may make it possible in the future to swap out lua versions in case you want to build this lib for your own game.